pyproject.toml
[project]
name = "edu-librarian"
version = "0.1.0"
description = "Simple chatbot with RAG"
readme = "README.md"
requires-python = ">=3.12,<=3.13"
dependencies = [
    "fastapi[standard]==0.135.3",
    "jinja2==3.1.6",
    "langchain==1.2.15",
    "langchain-ollama==1.1.0",
    "langchain-openai==1.1.12",
    "loguru==0.7.3",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
]


[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"
